Ascii_Art/requirements.txt
Pillow==10.2.0  # Python Imaging Library for image processing 